+++ csatar-plugins/csatar/models/LegalRelationship.php
@@ -57,4 +57,10 @@ class LegalRelationship extends Model
     {
         return $query->join('csatar_csatar_associations_legal_relationships', 'csatar_csatar_legal_relationships.id', '=', 'csatar_csatar_associations_legal_relationships.legal_relationship_id')->where('association_id', $id);
     }
+
+    public static function getInvalidDataId()
+    {
+        $data = self::where('name', 'Érvénytelen adat')->first();
+        return isset($data) ? $data->id : null;
+    }
 }